+++ projects/preload/src/app/app/app.module.spec.ts
@@ -1,7 +1,19 @@
+import { contextBridge } from 'electron';
 import { AppModule } from './app.module';
 
+jest.mock('electron', () => ({ contextBridge: { exposeInMainWorld: jest.fn() } }));
 describe('AppModule', () => {
+  let module: AppModule;
+  beforeEach(() => {
+    module = new AppModule();
+  });
   it('should create an instance', () => {
-    expect(new AppModule()).toBeTruthy();
+    expect(module).toBeTruthy();
+  });
+  it('should have an api', () => {
+    expect(module.api).toBeTruthy();
+  });
+  it('should expose the api under `window.electron`', () => {
+    expect(contextBridge.exposeInMainWorld).toHaveBeenCalledWith('electron', module.api);
   });
 });