+++ projects/renderer/src/app/services/git/git.service.ts
@@ -8,7 +8,9 @@ import { WindowService } from '../window/window.service';
 })
 export class GitService {
 
-  constructor(private dialogService: DialogService, private windowService: WindowService) { }
+  constructor(private dialogService: DialogService, private windowService: WindowService) {
+    window.electron.git.onprogress(event => console.debug(event));
+  }
   isRepo = window.electron.git.isRepo;
   setup = window.electron.git.setup;
   updateSetup = window.electron.git.updateSetup;