+++ projects/shared/services/image/image.service.d.ts
@@ -7,6 +7,8 @@ export interface ImageService {
   getImageMap: () => Promise<ImageMap>;
   setImageMap: (imageMap: ImageMap) => Promise<void>;
   readImage: (path: string) => Promise<string>;
+  stringToBase64: (string: string | ArrayBuffer) => string;
+  getExtension: (path: string) => Promise<string>;
 }
 export type Image = { file: string, alt?: string; };
 export type ImageMap = { [directory: string]: Image[]; };